+++ src/apps/chat-lab/public/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to FIDU Chat Lab will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.10] - UNRELEASED
+
+### Added
+- **Last Sync UI**: Extra UI to show last successful sync to google drive, and current sync health status
+
+### Changes
+- **Drive Sync Resilience**: Added extra mechanisms to try to automatically recover from sync failures from long lived sessions. 
+- **Allow Multiple Context Selection**: Multiple contexts can now be added to a prompt at once. 
+
 ## [0.1.9] - 2025-12-04
 
 ### Added